refacor(모든 테스트): service 패키지로 리팩터링
